+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/entity/DocumentHolder.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2014 - 2021 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.testsuite.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import java.io.Serializable;
+
+/**
+ *
+ * @author Christian Beikov
+ * @since 1.6.0
+ */
+@Entity
+public class DocumentHolder implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    private Long id;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "document_id")
+    // CHECKSTYLE:OFF: MemberName
+    private Document Document;
+    // CHECKSTYLE:ON: MemberName
+
+    public DocumentHolder() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Document getDocument() {
+        return Document;
+    }
+
+    public void setDocument(Document document) {
+        this.Document = document;
+    }
+
+}